backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
contourpy==1.0.5
cycler==0.11.0
Django==4.1.1
django-cors-headers==3.13.0
django-json==0.27
django-rest-auth==0.9.5
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
fonttools==4.37.3
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pip==22.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
setuptools==63.4.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.2
wheel==0.37.1
wincertstore==0.2
SQLAlchemy==1.4.41
PyMySQL==1.0.2
greenlet==1.1.3
